+++ FTAnalyzer.Shared/Displays/DisplayGreatStats.cs
@@ -3,10 +3,10 @@
     public class DisplayGreatStats
     {
         public string RelationToRoot { get; }
-        public long RelationSort { get; }
+        public decimal RelationSort { get; }
         public int Count { get; }
 
-        public DisplayGreatStats(string relation, long sort, int count)
+        public DisplayGreatStats(string relation, decimal sort, int count)
         {
             RelationToRoot = relation;
             RelationSort = sort;
